Further map functionality
Buggy stuff relating to variables not being properly saved, moved
testscreen to setup
