+++ digitalComponent/digitalComponent.pyde
@@ -12,8 +12,31 @@ def setup():
     noStroke()
     # Caladea = loadFont('Caladea-Regular-48.vlw')
     # textFont(Caladea)
-    constructTestScreen()
-    constructTestScreen2()
+    map1 = clickableMap(1000,500,500,300,'Map1',5,3)
+    typer = textInput(300,100,1000,100,'textIn1')
+    global map1
+    global typer
+    testScreen = Screen('testScreen')
+    screenManager.addScreen(testScreen)
+    testScreen.addItem(checkbox(200,200,40,20,'checkbox1',False,'kdfs'))
+    testScreen.addItem(linkButton(10,10,100,100,'linkButton1','testScreen2',screenManager))
+    testScreen.addItem(textBox(10,10,100,100,'buttonTextBox', ' Goto Screen 2', 20))
+    testScreen.addItem(textBox(300,300,1000,1000,'textBox1', 'Screen 1', 100))
+    testScreen.addItem(textBox(300,100,1000,100,'textBox3',typer.intext.tString,20))
+    testScreen.addItem(valButton(400,400,100,100,'val+',1,map1.columns,map1.rows))
+    map1.editTile(3,3,clickableDesert(0,0,0,0,''))
+    map1.editTile(3,2,clickableSwamp(0,0,0,0,''))
+    map1.editTile(1,1,clickableForest(0,0,0,0,''))
+    map1.editTile(5,2,clickableMountain(0,0,0,0,''))
+    map1.editTile(4,2,clickableHighland(0,0,0,0,''))
+    testScreen.addItem(map1)
+    screenManager.currentScreen = testScreen
+    testScreen2 = Screen('testScreen2')
+    screenManager.addScreen(testScreen2)
+    testScreen2.addItem(linkButton(10,10,100,100,'linkButton2','testScreen',screenManager))
+    testScreen2.addItem(textBox(10,10,100,100,'buttonTextBox', ' Goto Screen 1', 20))
+    testScreen2.addItem(textBox(300,300,1000,1000,'textBox2', 'Screen 2', 100))
+    testScreen2.addItem(typer)
